requirements.txt
absl-py==0.11.0
apipkg==1.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
distro==1.5.0
execnet==1.8.0
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.0
mock==4.0.3
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
parso==0.8.1
path==15.1.0
path.py==12.5.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.14
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.2
pytest-shutil==1.7.0
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-build==0.11.1
scikit-image==0.18.1
scipy==1.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
tifffile==2021.1.14
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
